% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_googleDriveConnections.R
\name{create_new_sheet}
\alias{create_new_sheet}
\title{Create a Google Sheet.}
\usage{
create_new_sheet(folder_id, sheet_name, template)
}
\arguments{
\item{folder_id}{Folder ID.}

\item{sheet_name}{Sheet name.}

\item{template}{Template.}
}
\value{
A Google Sheet.
}
\description{
Create a Google Sheet.
}
