% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_customValueBox.R
\name{customValueBox}
\alias{customValueBox}
\title{Custom Vale Box.}
\usage{
customValueBox(value, name, change, unit = "", show_change = FALSE)
}
\arguments{
\item{value}{The value to be displayed in the value box.}

\item{name}{The name of the value box.}

\item{change}{The change in the value.}

\item{unit}{The unit of the value.}

\item{show_change}{Whether to show the change in the value box.}
}
\value{
A custom value box.
}
\description{
Custom Vale Box.
}
\examples{
customValueBox(100, "Test", 10, "kg", show_change = TRUE)
}
