% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_templateDataEntrySpreadSheet.R
\name{farm_template_gs}
\alias{farm_template_gs}
\title{farm_template_gs Function to create the spreasheet.}
\usage{
farm_template_gs()
}
\value{
A list with the template of the spreadsheet.
}
\description{
farm_template_gs Function to create the spreasheet.
}
