% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_reportHeader.R
\name{generate_logo_header}
\alias{generate_logo_header}
\title{Generate Logo Header of Quarto Report.}
\usage{
generate_logo_header(consultoria, logo)
}
\arguments{
\item{consultoria}{Consulting name.}

\item{logo}{Path to the logo.}
}
\value{
A header with the consulting name and logo.
}
\description{
Generate Logo Header of Quarto Report.
}
