% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_reportHeader.R
\name{generate_text_header}
\alias{generate_text_header}
\title{Generate Header of Quarto Report.}
\usage{
generate_text_header(propriedade, fazendeiro, localidade, tecnico, data)
}
\arguments{
\item{propriedade}{Farm name.}

\item{fazendeiro}{Farmer name.}

\item{localidade}{Location.}

\item{tecnico}{Technician.}

\item{data}{Visit date.}
}
\value{
A header for the Quarto report.
}
\description{
Generate Header of Quarto Report.
}
