% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_googleDriveConnections.R
\name{list_farms}
\alias{list_farms}
\title{List the farms in a folder.}
\usage{
list_farms(folder_id)
}
\arguments{
\item{folder_id}{Folder ID.}
}
\value{
A list of farms.
}
\description{
List the farms in a folder.
}
